# core: async hash calculation