--- a/packages/hydrooj/src/lib/hash.hydro.ts
+++ b/packages/hydrooj/src/lib/hash.hydro.ts
@@ -1,7 +1,12 @@
 import crypto from 'crypto';
 
-function hash(password: string, salt: string) {
-    return crypto.pbkdf2Sync(password, salt, 100000, 64, 'sha256').toString('hex').substring(0, 64);
+async function hash(password: string, salt: string): Promise<string> {
+    return new Promise((resolve, reject) => {
+        crypto.pbkdf2(password, salt, 100000, 64, 'sha256', (err, key) => {
+            if (err) reject(err);
+            else resolve(key.toString('hex').substring(0, 64));
+        });
+    });
 }
 
 export = hash;