--- a/packages/hydrooj/src/lib/hash.hydro.ts
+++ b/packages/hydrooj/src/lib/hash.hydro.ts
@@ -1,6 +1,6 @@
 import crypto from 'crypto';
 
-async function hash(password: string, salt: string): Promise<string> {
+function hash(password: string, salt: string): Promise<string> {
     return new Promise((resolve, reject) => {
         crypto.pbkdf2(password, salt, 100000, 64, 'sha256', (err, key) => {
             if (err) reject(err);