The `pbkdf2` function is a CPU consuming operation (~50ms on m1) so using `sync` version blocks the Nodejs event loop for a while during request. Changing it into async version allows it to be calculated in worker thread without blocking. Stress test shows ~75% latency reduction.

before:
```
$ wrk -c8 -t1 -d5s -s login.lua http://localhost:2333/login
Running 5s test @ http://localhost:2333/login
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   398.81ms  113.01ms 711.82ms   72.45%
    Req/Sec    19.38      6.48    34.00     79.17%
  98 requests in 5.01s, 40.96KB read
Requests/sec:     19.55
Transfer/sec:      8.17KB
```

after:
```
$ wrk -c8 -t1 -d5s -s login.lua http://localhost:2333/login
Running 5s test @ http://localhost:2333/login
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   101.32ms    4.36ms 141.85ms   94.43%
    Req/Sec    78.42      7.85    90.00     74.00%
  395 requests in 5.05s, 165.10KB read
Requests/sec:     78.24
Transfer/sec:     32.70KB

```