@criyle 2024-04-08
> LGTM, but you may need edit @hydrooj/migrate too, then you can direct await in `result = await h(password, this._salt, this);`.
> 
> See https://github.com/hydro-dev/Hydro/blob/master/packages/migrate/index.ts#L62 .

I don't want to break the existing plugin system by this change since there might be other implementation of customized plugin defines their own hash functions or someone might forget to upgrade `migration` plugins with `hydrooj`. At the same time, I believe the current implementation just extend the current functionality to support async hash functions without a breaking change.